ecommerce_api/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
bcrypt
pyjwt
email-validator